feat(tools): add process_chunks.py script and fix database imports (T028)
- Add process_chunks.py script for processing chunk manifests into database
- Fix database import structure by moving database.py into database/__init__.py
- Resolves relative import issues that prevented standalone script execution
- Script supports parallel processing, resume capability, and status monitoring
- Includes comprehensive CLI with chunk ID filtering and cleanup options
- Mark T028 complete in tasks.md
